requirements.txt
Django>=4.2,<5.0
whitenoise>=6.5.0
PyJWT==2.8.0
dotenv